### 2023-07-10 #120 @JacksonBurns
`astartes`
Submitting Author: Name (@JacksonBurns)
All current maintainers: (@kspieks, @himaghna)
Package Name: `astartes`
One-Line Description of Package: Better Data Splits for Machine Learning
Repository Link:  https://github.com/JacksonBurns/astartes
Version submitted:   v[1.1.2](https://github.com/JacksonBurns/astartes/releases/tag/v1.1.2)
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8431894.svg)](https://doi.org/10.5281/zenodo.8431894)
Version accepted: v[1.1.3](https://github.com/JacksonBurns/astartes/releases/tag/v1.1.3)

---

_note: this is a selection from the abstract of the JOSS paper_

Machine Learning (ML) has become an increasingly popular tool to accelerate traditional workflows. Critical to the use of ML is the process of splitting datasets into training, validation, and testing subsets that are used to develop and evaluate models. Common practice in the literature is to assign these subsets randomly. Although this approach is fast and efficient, it only measures a model's capacity to interpolate. Testing errors from random splits may be overly optimistic if given new data that is dissimilar to the scope of the training set; thus, there is a growing need to easily measure performance for extrapolation tasks. To address this issue, we report astartes, an open-source Python package that implements many similarity- and distance-based algorithms to partition data into more challenging splits. Separate from astartes, users can then use these splits to better assess out-of-sample performance with any ML model of choice.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is data scientists, machine learning scientists, and domain scientists using machine learning. The applications of `astartes` include rigorous ML model validation, automated featurization of chemical data (with flexibility to add others, and instructions for doing so), and reproducibility.

We position `astartes` as a replacement to `scikit-learn`'s provides `train_test_split` function, but with greater flexibility for sampling algorithms, and availability of `train_val_test_split` for more rigorous validation.

N/A

__on a separate `joss-paper` branch__
- [x] The package is deposited in a long-term repository with the DOI: [10.5281/zenodo.7884532](https://zenodo.org/record/7884532)
