### 2023-07-11 #120 @cmarmo
Hi @JacksonBurns,
Thanks for submitting `astartes` to pyOpenSci and welcome to our submission procedure!
I'm Chiara and I will take care of the process as Editor.

After checking with the editorial board, I'm glad to confirm that  your package is considered in scope with a pyOpenSci submission.
I will perform the editor checks by the end of the week, and start looking for reviewers.

In the meanwhile I have a first question for you, if you don't mind.
While the statistical side of the package is clear and answers to a well known need of the scikit-learn user community (see eg https://github.com/scikit-learn/scikit-learn/issues/13990), I would like to read a bit more about the package connections with chemical engineering. I guess your scientific application was the motivation for this development in the first place and in the documentation the two aspects look quite unrelated. Perhaps a small pedagogical section about chemical applications in the documentation will also be helpful? Thanks!
