### 2023-07-12 JacksonBurns/astartes#144 @JacksonBurns
pyOpenSci Initial Review Cleanup
## Feedback from Initial Submission to pyOpenSci

In the meanwhile I have a first question for you, if you don't mind.
While the statistical side of the package is clear and answers to a well known need of the scikit-learn user community, I would like to read a bit more about the package connections with chemical engineering. I guess your scientific application was the motivation for this development in the first place and in the documentation the two aspects look quite unrelated. Perhaps a small pedagogical section about chemical applications in the documentation will also be helpful? Thanks!

_Originally posted by @cmarmo in https://github.com/pyOpenSci/software-submission/issues/120#issuecomment-1631493213_

We need to clarify the applications of `astartes` to chemical engineering _in particular_ rather than just the importance in terms of statistics and ML, broadly speaking. As mentioned by @kspieks offline there is some content in the JOSS paper that could be pulled from to flesh this out.

I would suggest that we make a separate markdown document, like the 'sklearn to astartes' one and reference it in the main README.
