### 2023-07-14 #120 @JacksonBurns
Hi @cmarmo thanks for the speedy work!

> * [ ]  **Version** (JOSS only, may be post-review): Does the release version given match the GitHub release (v1.0.0)?

I have just turned on the Zenodo automatic packaging releases (which I also just learned was a thing) so hopefully this doesn't fall behind again! I have also uploaded the current version (1.1.1) to Zenodo manually.

> The package is already in excellent shape, congratulations and thanks to all the co-authors for your hard work! I have some comments, they are not meant to be blockers for the review process, just suggestions for possible improvements.

Thank you for the kind words!

> * It seems to me that the [introduction about splitting algorithms](https://jacksonburns.github.io/astartes/README.html#rational-splitting-algorithms) is missing some references for those, like me, not really into chemical engineering. In particular I looked for the original work defining the Kennard-Stone algorithm (Kennard, R.W., and Stone, L.A., 1969. Computer aided design of experiments. Technometrics 11, 137-148) and I would like to see it cited in your documentation. The other algorithms also would benefit from some more references, when possible openly accessible, as unfortunately the links to most of the articles are pay-walled.

The reference for the Kennard-Stone algorithm is in the [Implemented Sampling Algorithms table](https://github.com/JacksonBurns/astartes#implemented-sampling-algorithms) but it is not at all obvious what the links in the `References` column actually are. We will clarify this, as well as add references where missing.

As far as accessible articles, we can also try and find open-access versions of those which are pay-walled.

> * Some external links are not correctly rendered ... I'm afraid I couldn't find a solution for backquoted text with external
>   link in sphinx... but I think having the link rendered is more readable. (see for example
>   `` ``AIMSim` <https://vlachosgroup.github.io/AIMSim/README.html>`_ ``  can be replaced by ``  `AIMSim <https://vlachosgroup.github.io/AIMSim/README.html>`_ ``)

:+1: will do.

> * Having jupyter notebooks in the documentation helps the user to understand how to use the package: however, I wasn't able to run the binder links (because of the initialization failing with the error "no space left on device").
>   I know Binder is downsizing, if you don't have alternatives perhaps it would be safer to just remove the links, in order to not unnecessarily pressure the system.

We can definitely do that. We will perhaps try to migrate these to target Google Colab.

> * The `CONTRIBUTING.md` file is clear and complete. As users are explicitly asked to provide the package version when opening an issue, I'm suggesting to define an `astartes.__version__` attribute and to detaile how to obtain the information when debugging.

:+1: sounds good!

> * The API section of the documentation would benefit from more examples: the user discovers in it that the package contains a lot more than the train-test-split functions, but it is not always clear how to exploit all this tools.

:+1: will do.

> I think this package may be relevant for the scikit-learn user community, perhaps you might be interested in linking `astartes` in the [related project page](https://scikit-learn.org/stable/related_projects.html#other-estimators-and-tasks) among the preprocessor tools, once the review process has come to an end.

Didn't know this existed, thanks for sharing! We will do that.

> I'm going to look for the two reviewers now (good news, I already found one! :) ). I will be back to you in a week (hopefully sooner), to check if we are able to start the process.

:tada:
