### 2023-07-14 JacksonBurns/astartes#144 @JacksonBurns
## Additional Comments:
(see [here](https://github.com/pyOpenSci/software-submission/issues/120#issuecomment-1635069583) for full explanations)

### References

Many references in the README are paywalled, try and find open-access versions. Some references are also missing, include them. Also make sure that the `References` column of the `Included Sampling Algorithms` table clarifies what the link actually is a reference to (documentation, a paper, etc.).

### Formatting Fixes

External links with backquotes are breaking in the Sphinx rendered documentation.

### Documentation

Add further examples of some of the backend, and how users could leverage other functionality implemented in `astartes` besides just `train_(val_)test_split`.

### Version Attribute

Specify a `__version__` attribute, as well as directions for how to retrieve it. Can use [this](https://packaging.python.org/en/latest/guides/single-sourcing-package-version/) to keep compatibility with `pyproject.toml`:
```toml
[project]
name = "astartes"
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {attr = "package.__version__"}
```
Which is compatible with both `from importlib.metadata import version` and `astartes.__version__` approaches.
