### 2023-07-14 53490ab
reorganize, add to README for domain applications
