### 2023-07-14 edbb407
add brief description of chemical engineering/chemistry applications
