### 2023-07-14 cc76ed6
move version attribute, improve CONTRIBUTING with version and dev notes
