+++ pyproject.toml
@@ -4,7 +4,6 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "astartes"
-version = "1.1.1"
 authors = [
     { name = "Jackson Burns", email = "jwburns@mit.edu" },
     { name = "Himaghna Bhattacharjee", email = "himaghna@udel.edu" },
@@ -22,6 +21,10 @@ classifiers = [
 urls = { Homepage = "https://github.com/JacksonBurns/astartes" }
 requires-python = ">=3.7"
 dependencies = ["scikit_learn", "tabulate", "numpy", "scipy", "pandas"]
+dynamic = ["version"]
+
+[tool.setuptools.dynamic]
+version = { attr = "package.__version__" }
 
 [project.optional-dependencies]
 molecules = ["aimsim"]
