### 2023-07-14 ead11fb
fix incorrectly rendering links, update citation
