### 2023-07-14 5ca34c2
move all of the demo notebooks to google colab instead of binder
response to an editor comment that they would not load on binder, they seem to work fine on google (with these changes)
