### 2023-07-14 JacksonBurns/astartes#144 @JacksonBurns
All of this is done in #147 except for fixing the references and some documentation for the internal functions.
