### 2023-07-14 #120 @JacksonBurns
@cmarmo quick question - we are getting these suggested changes made now (https://github.com/JacksonBurns/astartes/pull/147). Since we submitted `astartes v1.1.1`, can we go ahead and merge into `main`, or will that interfere with the review process?
