### 2023-07-14 #120 @cmarmo
> Since we submitted `astartes v1.1.1`, can we go ahead and merge into `main`, or will that interfere with the review process?

Would you be ok with making a quick release of the changes before the review starts and update the description? As far as the version does not change during the review process I think this is not an issue. Thanks for addressing the comments already!
