### 2023-07-16 c9726fe
improve documentation of internal functions, developer guidelines
