### 2023-07-17 03d0376
Minor: fix typo and update wording
