### 2023-07-17 29aef8d
patch fixes for the previous patch fixes
