+++ .github/workflows/gen_docs.yml
@@ -46,7 +46,7 @@ jobs:
           m2r README.md
           m2r CONTRIBUTING.md
           m2r sklearn_to_astartes.md
-          mv README.rst 
+          mv README.rst docs
           mv CONTRIBUTING.rst docs
           mv sklearn_to_astartes.rst docs
           cd docs
