### 2023-07-20 #120 @cmarmo
Hi @JacksonBurns , thank you for your new release: I've still spotted some small issues in the documentation, but I think it's better to move forward with the review.
I'm stll looking for the second reviewer. I hope to find someone by the end of the week. Thank you for your patience.
