### 2023-07-26 e6eb887
add a `_before_sample` method to clarify input validation in samplers
right now we have some samplers that have to check for valid inputs with things like if arg[0] which is confusing this slight refactor changes that
