+++ astartes/samplers/extrapolation/scaffold.py
@@ -28,14 +28,13 @@ from astartes.utils.warnings import NoMatchingScaffold
 
 class Scaffold(AbstractSampler):
     def __init__(self, *args):
-        # ensure that X (i.e. args[0]) contains entries that are either a SMILES string or an RDKit Molecule
-        if not isinstance(args[0][0], str) and not isinstance(
-            args[0][0], Chem.rdchem.Mol
-        ):
+        super().__init__(*args)
+
+    def _before_sample(self):
+        # ensure that X contains entries that are either a SMILES string or an RDKit Molecule
+        if not all(isinstance(i, str) for i in self.X) and not all(isinstance(i, Chem.rdchem.Mol) for i in self.X):
             msg = "Scaffold class requires input X to be an iterable of SMILES strings"
             raise TypeError(msg)
-
-        super().__init__(*args)
 
     def _sample(self):
         """Implements the Scaffold sampler to identify clusters via a molecule's Bemis-Murcko scaffold."""
@@ -46,8 +45,7 @@ class Scaffold(AbstractSampler):
         for cluster_id, (scaffold, indices) in enumerate(scaffold_to_indices.items()):
             if scaffold == "":
                 warnings.warn(
-                    f"No matching scaffold was found for the {len(indices)} "
-                    f"molecules corresponding to indices {indices}",
+                    f"No matching scaffold was found for the {len(indices)} " f"molecules corresponding to indices {indices}",
                     NoMatchingScaffold,
                 )
             for idx in indices:
@@ -67,9 +65,7 @@ class Scaffold(AbstractSampler):
         """
         scaffolds = defaultdict(set)
         for i, mol in enumerate(mols):
-            scaffold = self.generate_bemis_murcko_scaffold(
-                mol, self.get_config("include_chirality", False)
-            )
+            scaffold = self.generate_bemis_murcko_scaffold(mol, self.get_config("include_chirality", False))
             scaffolds[scaffold].add(i)
 
         return scaffolds
@@ -109,13 +105,7 @@ class Scaffold(AbstractSampler):
         Returns:
             Bemis-Murcko scaffold
         """
-        mol = (
-            self.str_to_mol(mol)
-            if isinstance(mol, str)
-            else mol
-        )
-        scaffold = MurckoScaffold.MurckoScaffoldSmiles(
-            mol=mol, includeChirality=include_chirality
-        )
+        mol = self.str_to_mol(mol) if isinstance(mol, str) else mol
+        scaffold = MurckoScaffold.MurckoScaffoldSmiles(mol=mol, includeChirality=include_chirality)
 
         return scaffold
