### 2023-07-26 13417d9
`_before_sample` should be concrete, but simply pass
