### 2023-07-26 c0afc69
make the kennard stone algorithm an importable util, refactor ks + spxy
