### 2023-07-27 7119194
refactor fast_kennard_stone for performance improvements, reusablility
removed an unnecessary operation in the for loop, slightly changed the implementation for better readability
