+++ astartes/samplers/interpolation/kennardstone.py
@@ -1,5 +1,6 @@
 from astartes.samplers import AbstractSampler
 from astartes.utils.fast_kennard_stone import fast_kennard_stone
+from scipy.spatial.distance import pdist, squareform
 
 
 class KennardStone(AbstractSampler):
@@ -10,4 +11,11 @@ class KennardStone(AbstractSampler):
         """
         Implements the Kennard-Stone algorithm
         """
-        self._samples_idxs = fast_kennard_stone(self.X, self.get_config("metric", "euclidean"))
+        self._samples_idxs = fast_kennard_stone(
+            squareform(
+                pdist(
+                    self.X,
+                    self.get_config("metric", "euclidean"),
+                )
+            )
+        )
