### 2023-08-01 #120 @cmarmo
Hello @JacksonBurns , I'm finally back! Thank you for your patience!
Sorry it took more time than expected.

@BerylKanali, @du-phan welcome! :wave:
Thank you for volunteering to review for pyOpenSci!

Please take some time to introduce yourself here before starting the review.

- [ ] Beryl Kanali survey completed.
- [ ] Du Phan survey completed.

Note that we ask reviewers to complete reviews in three weeks, the review for `astartes` is then due for **August 22nd**, please let me know if you have any major issue preventing to meet this deadline.
Please get in touch with any questions or concerns!

Thanks again for your commitment!
