### 2023-08-07 #120 @du-phan
Hi team,
Happy to help with this review! I'm a research scientist at [Dataiku](https://www.dataiku.com/), a software company developing a data science platform. We are part of the scikit-learn consortium and have collaborated with them for quite some time. We also developed and maintained our own os packages for some specific ML use cases, for example [Cardinal](https://github.com/dataiku-research/cardinal) for active learning.
