### 2023-08-07 #120 @BerylKanali
Hi everyone,

My name is Beryl Kanali, I am a Data Scientist and Open Source Advocate. I am also currently a graduate student. I like contributing to open source and I am happy to be a reviewer for `astartes`.

@cmarmo I have completed the pre-review survey.
