### 2023-08-17 a0f5920
remove metapackage, do a two package install instead
