### 2023-08-24 #120 @du-phan
Hi Jackson,

Sorry that my review takes a while, august is always a slow month.

Thank you for your contribution! It is a nice package.

- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README:

Estimated hours spent reviewing: 4-5

---

#### Review Comments:
##### Overview
The `astartes` package tackles an important step in the Machine Learning pipeline that is usually overlooked: generating robust data for training and validating the model performance. It provides several strategies to go beyond the classical random sampling and thus allows data practitioner to evaluate their model under different scenarios of interpolation/extrapolation.

##### Usage
- The two main functions are `train_test_split()` and `train_val_test_split()`, which is intuitive enough for users as they just need to swap their sklean's function by those. That's great!
- However, the devil is in the details, depending on the kind of sampler used (extrapolative vs interpolative), the number of returned arrays is not the same (as there can be cluster ids that are returned too). As this point is not clear in the doc, users can get confused (I was).
- Some method does not support categorical features, which is understandable when understanding how the sampler works, but maybe we can add those details somewhere in the doc.
- Being able to use the samplers directly is a nice addition!
- The `generate_regression_results_dict` is a helpful util function, but it looks a bit "experimental" to me as 1) it just supports the regression model and 2) users cannot choose the desired metrics. It feels like with a bit of work we can have something more generic.

##### Documentation
- The statement of needs is lacking in the README, maybe you can add the one in the paper to it?
- In the doc, the difference between extrapolative and interpolative sampling is discussed quite briefly. It is by working through the `split_comparison` notebook that I understood the "weird" behaviors of `train_test_split()` that I observed above. Maybe we can make this information easier for users to find out.
