### 2023-08-31 ce1e821
add status badges, format as a table
