+++ README.md
@@ -86,18 +86,19 @@ X_train, X_val, X_test = train_val_test_split(X, sampler = 'sphere_exclusion')
 ```
 You can now train your model with `X_train`, optimize your model with `X_val`, and measure its performance with `X_test`.
 
-### Evaluate the Impact of Splitting Algorithms
+### Evaluate the Impact of Splitting Algorithms on Regression Models
 For data with many features it can be difficult to visualize how different sampling algorithms change the distribution of data into training, validation, and testing like we do in some of the demo notebooks.
 To aid in analyzing the impact of the algorithms, `astartes` provides `generate_regression_results_dict`.
-This function allows users to quickly evaluate the impact of different splitting techniques on any model supported by `sklearn`. All results are stored in a dictionary format and can be displayed in a neatly formatted table using the optional `print_results` argument.
+This function allows users to quickly evaluate the impact of different splitting techniques on any model supported by `sklearn`.
+All results are stored in a dictionary format and can be displayed in a neatly formatted table using the optional `print_results` argument.
 
 ```python
 from sklearn.svm import LinearSVR
 
-from astartes.utils import generate_regression_results_dict
+from astartes.utils import generate_regression_results_dict as grrd
 
 sklearn_model = LinearSVR()
-results_dict = generate_regression_results_dict(
+results_dict = grrd(
     sklearn_model,
     X,
     y,
@@ -110,6 +111,16 @@ MAE   1.41522   3.13435   2.17091
 RMSE  2.03062   3.73721   2.40041
 R2    0.90745   0.80787   0.78412
 
+```
+
+Additional metrics can be passed to `generate_regression_results_dict` via the `additional_metrics` argument, which should be a dictionary mapping the name of the metric (as a `string`) to the function itself, like this:
+
+```python
+from sklearn.metrics import mean_absolute_percentage_error
+
+add_met = {"mape": mean_absolute_percentage_error}
+
+grrd(sklearn_model, X, y, additional_metric=add_met)
 ```
 
 ### Access Sampling Algorithms Directly
