### 2023-08-31 5e48414
updates to `generate_regression_results_dict`, see extended description
 - clarified in the README that this is only intended for regression (we should add an equivalent for classification, with different default metrics, that would perhaps involve creating a new abstract function that both this function and the new one would call)
 - make more general by allowing users to pass in other metrics, or to include their own
