### 2023-08-31 3952713
clarify early in README that extrapolation returns more arrays
