### 2023-08-31 ceaaede
add a README section explaining using `astartes` with categorical data
