### 2023-08-31 #120 @JacksonBurns
Hi @du-phan thank you for the review!

> * However, the devil is in the details, depending on the kind of sampler used (extrapolative vs interpolative), the number of returned arrays is not the same (as there can be cluster ids that are returned too). As this point is not clear in the doc, users can get confused (I was).

We have clarified this in the README, and in doing so referenced the demonstration notebook which you mentioned was useful as an explainer.

> * Some method does not support categorical features, which is understandable when understanding how the sampler works, but maybe we can add those details somewhere in the doc.

:+1: added details to the docs!

> * The `generate_regression_results_dict` is a helpful util function, but it looks a bit "experimental" to me as 1) it just supports the regression model and 2) users cannot choose the desired metrics. It feels like with a bit of work we can have something more generic.

We have added the ability to pass custom metrics to the function, and have ideas about adding a more generic `generate_results_dict` function which would be wrapped by `generate_regression_results_dict` and `generate_classification_results_dict` - although we tend to focus on regression.

> * The statement of needs is lacking in the README, maybe you can add the one in the paper to it?

:+1: added!

The changes I have described here are in [this draft pull request on `astartes`](https://github.com/JacksonBurns/astartes/pull/153) if you would like to take a look to see if they satisfy your comments here. After the review from @BerylKanali we will convert this into a normal Pull Request to finish the review process.

Thanks again!
