### 2023-08-31 aad672d
fix example for categorical data
