### 2023-09-06 #120 @BerylKanali
Hi Jackson,

Thank you for your patience, I am now settled and here is my review.

Estimated hours spent reviewing: 4hours 40 minutes

---

As a person who has used sklearn so many times, I am impressed by the `astartes` package. Being able quickly analyze the impact of sampling algorithms and to evaluate a model even futher under different interpoolation/extrapolation cases is an important step that will help in machine learning.

- The statement of need is missing. A user should be able to know why they need the package.
However, I have gone through the draft pull request as well and I believe the need has been adrressed and I can see two clear needs which is good.
- I also think it is better to have a clear target audience, with that statement of need we should be able to answer if the package is useful in machine learning generally or only those applying it in chemical data work. This is mostly helpful to beginners when they are trying to know if the package is useful to them or not. That is not clear yet
- Any plans on adding vignettes?
- **Suggestion**: In future you can add small sections of installation in windows, MacOS and linux like what you have done with pip and conda in the documentation, it has proven to help especially in promoting inclusion of underrepresented persons whose first language is not english, It goes directly to what they might need reducing what they might have to read through.

The package installation was easy and as per the instructions. The functional and performance claims are satisfied.

**Usability**
The functions are easy to use and remember especially if you have used sklearn before. It is also great that
- Do all methods support categorical data? This has been addressed in the draft PR and it is easier now to know which methods support it and how to use them.
- Are there any suggestions or ongoing work to improve ` generate_regression_results_dict` to make it more generic. Also is there adequate information on what it can presently do?
- From the draft PR I can see the issue on number of returned arrays as mentioned by @du-phan.
