### 2023-09-11 #120 @ucodery
Hello @JacksonBurns, I have been helping out the reviewers and in doing so have looked over some of `astartes` and hope you will allow me to submit some additional extra comments even as someone who has not before used sklearn.

- the version metadata can be dynamically determined from a code variable, meaning that the version in the REPL and in package METADATA can never be out of sync ([instructions here](https://packaging.python.org/en/latest/guides/single-sourcing-package-version/#single-sourcing-the-package-version))
- more and more open source projects are providing a pronunciation guide at the top of their readme. IMO this would help as I don't know where the syllables break up.
- there are several places a subclass overrides a method to then only call `super`, adding a call frame for no real change in logic.
- almost all of your `__init__.py`s hoist their submodules' public interfaces. Although this does shorten the dotted namespaces a user must use, it also will force every module to be found and loaded into the Python interpreter before any import from your package can be used. Yours is not a very large package but this will slow down all of its imports and increase its memory footprint, and this for a package that, it does not seem to me, a user would ever need to use all available interpolation/extrapolation variants.
