### 2023-09-18 84fd13d
add target audience, usrse notebook, and some typo fixes
