### 2023-09-18 bbfea5b
expand quickstart example to a simple vignette
