### 2023-09-18 af4292b
reword the platform installation instructions, move to be more prominent
