### 2023-09-18 8aa3f3b
add more explanation to generate_regression_results_dict
