### 2023-09-18 #120 @JacksonBurns
> #### Review Comments
> As a person who has used sklearn so many times, I am impressed by the `astartes` package. Being able quickly analyze the impact of sampling algorithms and to evaluate a model even futher under different interpoolation/extrapolation cases is an important step that will help in machine learning.

Thanks for the kind words!

> **Documentation**
>
> * The statement of need is missing. A user should be able to know why they need the package.
>   However, I have gone through the draft pull request as well and I believe the need has been adrressed and I can see two clear needs which is good.

:+1:

> * I also think it is better to have a clear target audience, with that statement of need we should be able to answer if the package is useful in machine learning generally or only those applying it in chemical data work. This is mostly helpful to beginners when they are trying to know if the package is useful to them or not. That is not clear yet

Added!

> * Any plans on adding vignettes?

I've expanded our Quick Start example to be copy-pastable - I think as far as other examples, we have some great demo notebooks are are now trying to push those in multiple places in the readme.

> * **Suggestion**: In future you can add small sections of installation in windows, MacOS and linux like what you have done with pip and conda in the documentation, it has proven to help especially in promoting inclusion of underrepresented persons whose first language is not english, It goes directly to what they might need reducing what they might have to read through.

:+1: we moved the platform installation tip above the conda/pip/source installation section. The only difference between platforms is the inclusion of quotes around one argument.

> **Functionality** The package installation was easy and as per the instructions. The functional and performance claims are satisfied.
>
> **Usability** The functions are easy to use and remember especially if you have used sklearn before. It is also great that
>
> * Do all methods support categorical data? This has been addressed in the draft PR and it is easier now to know which methods support it and how to use them.
> * Are there any suggestions or ongoing work to improve ` generate_regression_results_dict` to make it more generic. Also is there adequate information on what it can presently do?

One plan we have internally is to make another `generate_classification_results_dict`, which would involve a small internal refactor to create a more general `generate_results_dict`. We are more familiar with regression, so we are reading on what metrics would be good defaults and then we will consider where to go.

For this submission, though, I have added some additional details clarifying what the function actually returns.

> * From the draft PR I can see the issue on number of returned arrays as mentioned by @du-phan.

:+1:

Thank you for the review! All of the changes I mentioned making above are on [this draft PR](https://github.com/JacksonBurns/astartes/pull/153) (the same you already found). Please let me know how they look.
