### 2023-09-18 f83e72d
single-source the package version
