+++ astartes/__init__.py
@@ -1,7 +1,6 @@
 # convenience import to enable 'from astartes import train_test_split'
 from .main import train_test_split, train_val_test_split
 
-# update this in pyproject.toml, too
 __version__ = "1.1.2"
 
 # DO NOT do this:
