### 2023-09-18 faba6f3
remove the unnecessary constructor overrides and MTSD (didn't get to it)
