### 2023-09-18 #120 @JacksonBurns
> Hello @JacksonBurns, I have been helping out the reviewers and in doing so have looked over some of `astartes` and hope you will allow me to submit some additional extra comments even as someone who has not before used sklearn.

The additional feedback is appreciated!

> * the version metadata can be dynamically determined from a code variable, meaning that the version in the REPL and in package METADATA can never be out of sync ([instructions here](https://packaging.python.org/en/latest/guides/single-sourcing-package-version/#single-sourcing-the-package-version))

Thanks for the link - just done now.

> * more and more open source projects are providing a pronunciation guide at the top of their readme. IMO this would help as I don't know where the syllables break up.

:+1: this has come up a lot, should have added sooner!

> * there are several places a subclass overrides a method to then only call `super`, adding a call frame for no real change in logic.

Thanks for catching this, just removed them.

> * almost all of your `__init__.py`s hoist their submodules' public interfaces. Although this does shorten the dotted namespaces a user must use, it also will force every module to be found and loaded into the Python interpreter before any import from your package can be used. Yours is not a very large package but this will slow down all of its imports and increase its memory footprint, and this for a package that, it does not seem to me, a user would ever need to use all available interpolation/extrapolation variants.

I didn't know that incurred a cost like that, thanks for explaining. Since we are using a `SamplerFactory` (in `astartes/utils/sampler_factory.py`) which imports all of the samplers, would we still get the performance benefits of reworking the import statements? Is the extra cost only caused when the imports are in the `__init__.py`s?
