### 2023-09-19 #120 @cmarmo
@JacksonBurns thank you for your work and the answers provided to the reviewers.
As we are still in the three weeks window for answering the reviews, I was wondering if it would be too much work for you to split the [draft PR](https://github.com/JacksonBurns/astartes/pull/153) in multiple pull requests addressing specific items in the review. The comments for each commit are indeed clear, but there are 19 modified files and 13 commits not always separable. Having smaller and focused pull requests will simplify the work of the reviewers looking specifically to the modifications addressing their comments.
Do you think this is doable in a week?
Thank you very much for your time!
