### 2023-09-22 fce22a2
add a README section explaining using `astartes` with categorical data
