### 2023-09-22 f191c7d
add target audience, usrse notebook, and some typo fixes
