### 2023-09-22 08c1491
add more explanation to generate_regression_results_dict
