### 2023-09-22 d2c54fb
remove the unnecessary constructor overrides and MTSD (didn't get to it)
