+++ astartes/samplers/interpolation/__init__.py
@@ -1,4 +1,3 @@
 from .kennardstone import KennardStone
-from .mtsd import MTSD
 from .random_split import Random
 from .spxy import SPXY
