### 2023-10-08 #120 @cmarmo
Hello @du-phan and @BerylKanali , do you mind letting us know if Jackson addressed all your concerns in https://github.com/JacksonBurns/astartes/pull/154 and https://github.com/JacksonBurns/astartes/pull/155?
Thank you so much for yur collaboration! :pray:
