### 2023-10-11 #120 @BerylKanali
Hi @cmarmo  and @JacksonBurns thanks for the reminder.
I have looked at both PR's.
**Documentation**
Good job on addressing the concerns and naming the commits in a way that is easy to identify what was being worked on. I have also gone through the README and the documentation page.
I believe all my concerns have been addressed.

**Code**
I have gone through the commits and the issues they address and I believe you have addressed and explained everything well. I have no additional concerns on the code or functionality.

Good job and congratulations on a job well done!
