### 2023-10-11 3f58ac3
remove release check and conda build (on conda-forge now)
