+++ paper.bib
@@ -59,6 +59,17 @@
   publisher={ACS Publications}
 }
 
+@article{spiekermann2023comment,
+  title={Comment on `Physics-based representations for machine learning properties of chemical reactions'},
+  author={Spiekermann, Kevin A. and Stuyver, Thijs and Pattanaik, Lagnajit and Green, William H.},
+  journal={Machine Learning: Science & Technology},
+  volume={4},
+  number={4},
+  pages={048001},
+  year={2023},
+  publisher={IOP Publishing}
+}
+
 %%%%%%%%%% Datasets %%%%%%%%%%
 % Original QM9
 @article{ramakrishnan2014quantum,
