### 2023-10-15 #120 @cmarmo
🎉 `astartes` has been approved by pyOpenSci! Thank you for your submission and many thanks to @BerylKanali and @du-phan for reviewing this package! 😸
Maintainer and reviewers, do you mind filling out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7)?

The actions needed to finalize this review are detailed below.

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

---
