### 2023-10-16 #120 @JacksonBurns
> Dear @JacksonBurns,
>
> 🎉 `astartes` has been approved by pyOpenSci! Thank you for your submission and many thanks to @BerylKanali and @du-phan for reviewing this package! 😸 Maintainer and reviewers, do you mind filling out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7)?

🎉!  @kspieks, @himaghna, and I are filling out the survey!

>     * [ ]  Activate [Zenodo](https://zenodo.org/) watching the repo if you haven't already done so.

Already done :+1:

>     * [ ]  Tag and create a release to create a Zenodo version and DOI.
>
>     * [ ]  Add the badge for pyOpenSci peer-review to the README.md of . The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

Will do!

> It looks like you would like to submit this package to JOSS. Here are the next steps:
>
>     * [ ]   Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

Will do!

>     * [ ]  Login to the JOSS website and fill out the JOSS submission form using your Zenodo DOI. **When you fill out the form, be sure to mention and link to the approved pyOpenSci review.** JOSS will tag your package for expedited review if it is already pyOpenSci approved.

:+1:

I will keep this issue up-to-date as the reviews progress.

Thanks again to @cmarmo, @BerylKanali, and @du-phan for volunteering your time!
