### 2023-10-31 df4df7a
fix missing dois in bibliography
see this comment from the JOSS editorial bot: https://github.com/openjournals/joss-reviews/issues/5996#issuecomment-1783659584
