+++ paper.bib
@@ -56,6 +56,7 @@
   number={12},
   pages={4954--4965},
   year={2020},
+  doi={10.1021/acs.chemmater.0c01907.s001},
   publisher={ACS Publications}
 }
 
@@ -80,6 +81,7 @@
   number={1},
   pages={1--7},
   year={2014},
+  doi={10.1038/sdata.2014.22},
   publisher={Nature Publishing Group}
 }
 
@@ -91,6 +93,7 @@
   number={11},
   pages={2864--2875},
   year={2012},
+  doi={10.1021/ci300415d},
   publisher={ACS Publications}
 }
 
@@ -102,6 +105,7 @@
   number={1},
   pages={1--12},
   year={2022},
+  doi={10.1038/s41597-022-01529-6},
   publisher={Nature Publishing Group}
 }
 
@@ -125,6 +129,7 @@
   number={25},
   pages={3976--3986},
   year={2022},
+  doi={10.1021/acs.jpca.2c02614},
   publisher={ACS Publications}
 }
 
@@ -176,6 +181,7 @@
   year={2022},
   publisher={Royal Society of Chemistry},
   volume={24},
+  doi={10.1039/d1cp04422b},
   pages={2692--2705},
 }
 
@@ -203,6 +209,7 @@
   volume={3},
   number={5},
   pages={819--825},
+  doi={10.1039/d1cp04422b},
   year={2018},
   publisher={Royal Society of Chemistry}
 }
@@ -217,6 +224,7 @@
   volume={1},
   pages={763--778},
   year={2022},
+  doi={10.1039/d2dd00039c},
   publisher={Royal Society of Chemistry}
 }
 
@@ -228,6 +236,7 @@
   title={Construction of Balanced, Chemically Dissimilar Training, Validation and Test Sets for Machine Learning on Molecular Datasets},
   author={Tricarico, Giovanni A. and Hofmans, Johan and Lenselink, Eelke B. and Ramos, Miriam L{\'o}pez and Dr{\'e}anic, Marie-Pierre and Stouten, Pieter FW},
   year={2022},
+  doi={10.26434/chemrxiv-2022-m8l33-v2},
   journal={10.26434/chemrxiv-2022-m8l33}
 }
 
@@ -241,6 +250,7 @@
   volume={14},
   pages={1419--1433},
   year={2023},
+  doi={10.1039/d2sc06150c},
   publisher={Royal Society of Chemistry}
 }
 
@@ -269,6 +279,7 @@
   number={8},
   pages={084104},
   year={2022},
+  doi={10.1063/5.0079574},
   publisher={AIP Publishing LLC}
 }
 
@@ -285,6 +296,7 @@
   number={6},
   pages={064105},
   year={2021},
+  doi={10.1063/5.0059742},
   publisher={AIP Publishing LLC}
 }
 % we developed a graph-based neural network architecture and applied it to the problem of predicting the viscosity of binary liquid mixtures as a function of composition and temperature
@@ -297,7 +309,8 @@
   journal={Chem. Eng. J.},
   pages={142454},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.2139/ssrn.4289793}
 }
 
 % With this leave-one-reaction-out validation approach, we observed a MAE of 1.00 kcal mol−1 for GPRM3/2 (compared to 0.80 kcal mol−1 from normal cross-validation)
@@ -307,6 +320,7 @@
   author={Jorner, Kjell and Brinck, Tore and Norrby, Per-Ola and Buttar, David},
   journal={Chem. Sci.},
   volume={12},
+  doi={10.26434/chemrxiv.12758498},
   number={3},
   pages={1163--1175},
   year={2021},
[…]
