### 2023-11-05 a8a91d3
update readme and cff with new JOSS citation
