### 2023-11-06 c5467a7
run reproduce paper monthly instead of weekly
