+++ .github/workflows/reproduce_paper.yml
@@ -6,8 +6,6 @@ name: Reproduce Paper
 on:
   schedule:
     - cron: "0 8 3 * *"
-  push:
-    branches: [main]
   workflow_dispatch:
 
 concurrency:
