### 2023-11-12 6849c6f
add python 3.12 to ci, pyproject, readme - bump patch version 4 release
