### 2023-11-12 #120 @cmarmo
@JacksonBurns congratulations! astartes has been accepted on JOSS!
I have announced it "officially" on our [Discourse](https://pyopensci.discourse.group/t/welcome-astartes-to-the-pyopensci-ecosystem/394) too... ;)
If you and and your colleagues maintainers have been happy with the pyOpenSci review process fell free to let the world know via a blog or any kind of testimonial ... :).

@BerylKanali and @du-phan thank you once more for your work as reviewers: if you have some time to fill the [pos-review-survey](https://forms.gle/BLGVCfUdiJHS5YJY7) this will help us a lot.

I am going to close the issue now.
Let's go celebrate ! :partying_face:
